Move ToDiagnostic to a separate module
