+++ src/Control/Effect/Diagnostics.hs
@@ -45,9 +45,9 @@ module Control.Effect.Diagnostics (
 import Control.Algebra as X
 import Control.Carrier.Stack
 import Control.Effect.Lift (Lift)
-import Control.Exception (IOException, SomeException (..))
+import Control.Exception (IOException)
 import Control.Exception.Extra (safeCatch)
-import Data.Aeson (ToJSON, object, toJSON, (.=))
+import Data.Diagnostic (SomeDiagnostic (SomeDiagnostic), ToDiagnostic, renderDiagnostic)
 import Data.List (intersperse)
 import Data.List.NonEmpty qualified as NE
 import Data.Maybe (catMaybes)
@@ -66,31 +66,6 @@ data DiagErr m k where
   Recover :: m a -> DiagErr m (Maybe a)
   ErrorBoundary :: m a -> DiagErr m (Either FailureBundle a)
   Rethrow :: FailureBundle -> DiagErr m a
-
--- | A class of diagnostic types that can be rendered in a user-friendly way
-class ToDiagnostic a where
-  renderDiagnostic :: a -> Doc AnsiStyle
-
-instance ToDiagnostic (Doc AnsiStyle) where
-  renderDiagnostic = id
-
-instance ToDiagnostic Text where
-  renderDiagnostic = pretty
-
-instance ToDiagnostic SomeException where
-  renderDiagnostic (SomeException exc) =
-    "An exception occurred: " <> pretty (show exc)
-
--- | An error with a ToDiagnostic instance and an associated stack trace
-data SomeDiagnostic where
-  SomeDiagnostic :: ToDiagnostic a => [Text] -> a -> SomeDiagnostic
-
-instance ToJSON SomeDiagnostic where
-  toJSON (SomeDiagnostic path cause) =
-    object
-      [ "errorPath" .= path
-      , "errorCause" .= show (renderDiagnostic cause)
-      ]
 
 -- | Analagous to @throwError@ from the error effect
 fatal :: (Has Diagnostics sig m, ToDiagnostic diag) => diag -> m a
